+++ kiel/src/main/java/me/ibrahimyilmaz/kiel/adapter/RecyclerViewAdapter.kt
@@ -4,6 +4,8 @@ import android.view.ViewGroup
 import androidx.annotation.MainThread
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
+import me.ibrahimyilmaz.kiel.core.RecyclerViewHolder
+import me.ibrahimyilmaz.kiel.core.RecyclerViewHolderManager
 
 class RecyclerViewAdapter<T : Any, VH : RecyclerViewHolder<T>> private constructor(
     private val recyclerViewHolderManager: RecyclerViewHolderManager<T, VH>,
@@ -52,6 +54,7 @@ class RecyclerViewAdapter<T : Any, VH : RecyclerViewHolder<T>> private construct
     }
 
     companion object {
+
         @JvmStatic
         operator fun <T : Any, VH : RecyclerViewHolder<T>> invoke(
             recyclerViewHolderManager: RecyclerViewHolderManager<T, VH>,
@@ -59,13 +62,6 @@ class RecyclerViewAdapter<T : Any, VH : RecyclerViewHolder<T>> private construct
         ) = diffUtilCallbackFactory?.let {
             RecyclerViewAdapter(recyclerViewHolderManager, it)
         } ?: RecyclerViewAdapter(recyclerViewHolderManager)
-
-        @JvmStatic
-        @Deprecated("s plural typo removed", ReplaceWith("adapterOf"), DeprecationLevel.ERROR)
-        inline fun <T : Any> adaptersOf(
-            function: RecyclerViewAdapterBuilder<T>.() -> Unit
-        ): RecyclerViewAdapter<T, RecyclerViewHolder<T>> =
-            RecyclerViewAdapterBuilder<T>().apply(function).build()
 
         @JvmStatic
         inline fun <T : Any> adapterOf(