+++ kiel/src/main/java/me/ibrahimyilmaz/kiel/core/OnViewHolderBound.kt
@@ -1,3 +1,3 @@
-package me.ibrahimyilmaz.kiel.adapter
+package me.ibrahimyilmaz.kiel.core
 
 typealias OnViewHolderBound<T, VH> = (VH, Int, T) -> Unit