+++ kiel/src/main/java/me/ibrahimyilmaz/kiel/listadapter/RecyclerViewItemCallback.kt
@@ -1,0 +1,18 @@
+package me.ibrahimyilmaz.kiel.listadapter
+
+import android.annotation.SuppressLint
+import androidx.recyclerview.widget.DiffUtil
+
+class RecyclerViewItemCallback<T : Any> : DiffUtil.ItemCallback<T>() {
+
+    override fun areItemsTheSame(
+        oldItem: T,
+        newItem: T
+    ) = oldItem === newItem
+
+    @SuppressLint("DiffUtilEquals")
+    override fun areContentsTheSame(
+        oldItem: T,
+        newItem: T
+    ) = oldItem == newItem
+}