+++ kiel/src/test/java/me/ibrahimyilmaz/kiel/adapter/RecyclerViewAdapterBuilderTest.kt
@@ -16,7 +16,7 @@ class RecyclerViewAdapterBuilderTest {
     }
 
     @Test
-    fun `Should build an instance of RecyclerViewAdapterBuilder`() {
+    fun `Should build an instance of RecyclerViewAdapter`() {
         // GIVEN
         recyclerViewAdapterBuilder.register {
             layoutResource { 1 }